requirements_gemini.txt
# StoryFi Bittensor Subnet Requirements (Google Gemini 版本)
# ========================================================

# Core Bittensor SDK
bittensor>=6.0.0

# Data validation and parsing
pydantic>=2.0.0

# Async support
aiohttp>=3.9.0

# Google Gemini (替代 OpenAI)
google-generativeai>=0.3.0

# Embedding models (optional)
sentence-transformers>=2.2.0

# Data processing
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging and monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
